chore: add world story
